move updatedwind to stencils
